removed backfill and backfills command from CLI

+++ src/lib/createCli.ts
@@ -1,15 +1,15 @@
-import { convertDateToTimestamp } from "../utils/index";
-import backfillCommand from "./commands/backfill";
-import backfillsCommand from "./commands/backfills";
+import {
+	backfill as backfillCommand,
+	backfills as backfillsCommand
+} from "@algotia/core";
 import listPairsCommand from "./commands/list-pairs";
 import { ListOptions, DeleteOptions } from "../types/commands";
 import { program } from "commander";
-import { bail } from "../utils/index";
+import { confirmDangerous, bail } from "../utils/index";
 const packageJson = require("../../package.json");
 
 // argument parsers
 const pInt = (str: string) => parseInt(str, 10);
-const pDate = (str: string) => convertDateToTimestamp(str);
 
 // should create an interface for this
 export default (bootData) => {
@@ -21,15 +21,13 @@ export default (bootData) => {
 		.description("backfill historical data")
 		.requiredOption(
 			"-s, --since <since>",
-			"Unix timestamp (ms) of time to retrieve records from",
-			pDate
+			"Unix timestamp (ms) of time to retrieve records from"
 		)
 		.requiredOption("-p, --pair <pair>", "Pair to retrieve records for")
 		.option("-P, --period <period>", "Timeframe to retrieve records for", "1m")
 		.option(
 			"-u, --until <until>",
 			"Unix timestamp (ms) of time to retrieve records to",
-			pDate,
 			// default argument is server time in MS
 			exchange.milliseconds()
 		)
@@ -37,43 +35,31 @@ export default (bootData) => {
 			"-l, --limit <limit>",
 			"Number of records to retrieve at one time",
 			pInt,
-			10
+			100
 		)
 		.option(
-			"-n, --collection-name <collectionName>",
+			"-n, --document-name <documentName>",
 			"name for database refrence",
 			undefined
 		)
 		.action(async (options) => {
 			try {
-				const {
-					since,
-					pair,
-					period,
-					until,
-					limit,
-					collectionName
-				}: {
-					since: number;
-					pair: string;
-					period: string;
-					until: number;
-					limit: number;
-					collectionName: string;
-				} = options;
+				const { verbose } = program;
+				const { since, pair, period, until, limit, documentName } = options;
 
 				const opts = {
-					since,
+					sinceInput: since,
+					untilInput: until,
 					pair,
 					period,
-					until,
 					recordLimit: limit,
-					name: collectionName
+					documentName: documentName,
+					verbose: verbose
 				};
 
 				await backfillCommand(exchange, opts);
 			} catch (err) {
-				bail(err);
+				return Promise.reject(new Error(err));
 			}
 		});
 
@@ -100,7 +86,7 @@ export default (bootData) => {
 					await backfillsCommand.listAll(backfillsOptions);
 				}
 			} catch (err) {
-				bail(err);
+				return Promise.reject(new Error(err));
 			}
 		});
 
@@ -115,13 +101,19 @@ export default (bootData) => {
 				const deleteOptions: DeleteOptions = {
 					verbose
 				};
-				if (documentName) {
-					await backfillsCommand.deleteOne(documentName, deleteOptions);
+
+				const proceed = await confirmDangerous();
+				if (proceed) {
+					if (documentName) {
+						await backfillsCommand.deleteOne(documentName, deleteOptions);
+					} else {
+						await backfillsCommand.deleteAll(deleteOptions);
+					}
 				} else {
-					await backfillsCommand.deleteAll(deleteOptions);
+					bail("Bailing out of deleting all documents.");
 				}
 			} catch (err) {
-				bail(err);
+				return Promise.reject(new Error(err));
 			}
 		});
 
@@ -130,10 +122,9 @@ export default (bootData) => {
 		.description(
 			"Lists all the valid trading pairs from the exchange in your configuration."
 		)
-		.option("-a, --all", "Print tickers with all information", false)
+		.option("-v, --verbose")
 		.action((options) => {
-			console.log("opts - ", options.all);
-			listPairsCommand(exchange, options.all);
+			listPairsCommand(exchange, options.verbose);
 		});
 
 	program.parse(process.argv);
